skal flytte innholdet til forsiden til et eget komponent
